Fix a missing scan id int -> long casting + add some methods for MSn XICs

+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/bb/IBlobReader.java
@@ -27,9 +27,9 @@ public interface IBlobReader {
 	/**
 	 * 
 	 * @param i index of scan starting at 1
-	 * @return int, the ID of the scan at the specified index in the blob
+	 * @return long, the ID of the scan at the specified index in the blob
 	 */
-	int getScanIdAt(int i);
+	long getScanIdAt(int i);
 
 	/**
 	 *
